requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.0
psycopg2-binary==2.9.6
